ray_cluster/ray_embeddings.py
"""
Ray Remote Functions für Embedding-Generierung
Verteilt Embedding-Generierung über mehrere Mac Nodes
"""

import ray
import logging
import base64
import io
from typing import List, Dict, Any
from PIL import Image
import numpy as np

# MLX Import
try:
    import mlx.core as mx
    import mlx.nn as nn
    MLX_AVAILABLE = True
except ImportError:
    MLX_AVAILABLE = False

logger = logging.getLogger(__name__)

# Ray Remote Function für Embedding-Generierung
@ray.remote(num_gpus=0.5, num_cpus=2)
def generate_embedding_remote(face_crop_bytes: bytes) -> List[float]:
    """
    Ray Remote Function für Embedding-Generierung auf einem Worker Node
    
    Args:
        face_crop_bytes: Gesichts-Crop als Bytes
        
    Returns:
        Embedding-Vektor als Liste
    """
    try:
        # Bild laden
        image = Image.open(io.BytesIO(face_crop_bytes))
        
        # MLX Device setzen
        if MLX_AVAILABLE:
            mx.set_default_device(mx.gpu)
        
        # TODO: Echte Embedding-Generierung mit ArcFace-MLX
        # Hier würde das ArcFace-MLX Model verwendet werden
        # Für jetzt: Fallback-Implementierung
        
        embedding_dim = 512
        embedding = np.random.normal(0, 1, embedding_dim).tolist()
        
        # Normalisierung
        embedding_array = np.array(embedding)
        norm = np.linalg.norm(embedding_array)
        if norm > 0:
            embedding = (embedding_array / norm).tolist()
        
        logger.debug(f"Embedding generiert (Dimension: {len(embedding)})")
        return embedding
        
    except Exception as e:
        logger.error(f"Fehler bei Embedding-Generierung: {e}", exc_info=True)
        return []


@ray.remote(num_gpus=0.5, num_cpus=2)
def generate_embeddings_batch_remote(face_crops: List[bytes]) -> List[List[float]]:
    """
    Ray Remote Function für Batch Embedding-Generierung
    
    Args:
        face_crops: Liste von Gesichts-Crops als Bytes
        
    Returns:
        Liste von Embedding-Vektoren
    """
    embeddings = []
    for crop in face_crops:
        embedding = generate_embedding_remote.remote(crop)
        embeddings.append(embedding)
    
    # Warte auf alle Ergebnisse
    return ray.get(embeddings)


# Ray Actor für persistente Embedding Modelle
@ray.remote(num_gpus=0.5, num_cpus=2)
class EmbeddingActor:
    """Ray Actor für persistente Embedding Modelle"""
    
    def __init__(self):
        self.model = None
        self.initialized = False
        self.embedding_dim = 512
        
        if MLX_AVAILABLE:
            mx.set_default_device(mx.gpu)
        
        # TODO: Model laden
        # self.model = load_arcface_mlx()
        # self.initialized = True
    
    def encode(self, face_crop_bytes: bytes) -> List[float]:
        """Generiert Embedding mit geladenem Model"""
        if not self.initialized:
            # Fallback
            embedding = np.random.normal(0, 1, self.embedding_dim).tolist()
            embedding_array = np.array(embedding)
            norm = np.linalg.norm(embedding_array)
            if norm > 0:
                return (embedding_array / norm).tolist()
            return embedding
        
        # Echte MLX-Implementierung würde hier stehen
        return []
    
    def encode_batch(self, face_crops: List[bytes]) -> List[List[float]]:
        """Batch Embedding-Generierung"""
        embeddings = []
        for crop in face_crops:
            embedding = self.encode(crop)
            embeddings.append(embedding)
        return embeddings
