services/face_detection_service.py
#!/usr/bin/env python3
"""
Face Detection Service (MLX-optimiert)
Port: 5001
"""

import logging
import base64
import io
from flask import Flask, request, jsonify
from PIL import Image
import numpy as np

# MLX Import (wird später implementiert, wenn Modelle verfügbar sind)
try:
    import mlx.core as mx
    import mlx.nn as nn
    MLX_AVAILABLE = True
except ImportError:
    MLX_AVAILABLE = False
    logging.warning("MLX nicht verfügbar, verwende Fallback")

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = Flask(__name__)


class FaceDetectionModel:
    """Face Detection Model (MLX-optimiert)"""
    
    def __init__(self):
        self.model = None
        self.initialized = False
        # TODO: Model laden wenn verfügbar
        # self._load_model()
    
    def _load_model(self):
        """Lädt das Face Detection Model"""
        # TODO: RetinaFace-MLX Model laden
        # self.model = load_retinaface_mlx()
        self.initialized = True
    
    def detect(self, image: Image.Image) -> list:
        """
        Erkennt Gesichter in einem Bild
        
        Args:
            image: PIL Image
            
        Returns:
            Liste von Gesichtern mit Bounding Boxes
        """
        # Fallback-Implementierung (für Testing)
        # TODO: Durch echte MLX-Implementierung ersetzen
        if not self.initialized:
            logger.warning("Model nicht geladen, verwende Fallback")
            # Simuliere Gesichtserkennung
            width, height = image.size
            return [{
                "box": [width * 0.2, height * 0.2, width * 0.3, height * 0.3],
                "confidence": 0.95,
                "landmarks": {
                    "left_eye": [width * 0.25, height * 0.25],
                    "right_eye": [width * 0.35, height * 0.25],
                    "nose": [width * 0.3, height * 0.3],
                    "mouth_left": [width * 0.25, height * 0.35],
                    "mouth_right": [width * 0.35, height * 0.35]
                }
            }]
        
        # Echte MLX-Implementierung würde hier stehen
        # img_array = np.array(image.resize((640, 640)))
        # img_tensor = mx.array(img_array.transpose(2, 0, 1))
        # detections = self.model(img_tensor)
        # return self._process_detections(detections)
        
        return []


# Globales Model
face_model = FaceDetectionModel()


@app.route("/health", methods=["GET"])
def health():
    """Health Check Endpoint"""
    return jsonify({
        "status": "healthy",
        "service": "face_detection",
        "mlx_available": MLX_AVAILABLE,
        "model_loaded": face_model.initialized
    })


@app.route("/detect", methods=["POST"])
def detect():
    """
    Face Detection Endpoint
    
    Request Body:
    {
        "image": "base64_encoded_image",
        "return_landmarks": true,
        "return_confidence": true
    }
    
    Response:
    {
        "faces": [
            {
                "box": [x, y, width, height],
                "confidence": 0.95,
                "landmarks": {...}
            }
        ]
    }
    """
    try:
        data = request.get_json()
        
        if not data or "image" not in data:
            return jsonify({"error": "Missing 'image' in request body"}), 400
        
        # Base64 dekodieren
        image_data = base64.b64decode(data["image"])
        image = Image.open(io.BytesIO(image_data))
        
        # Face Detection
        faces = face_model.detect(image)
        
        # Response formatieren
        result = {"faces": []}
        for face in faces:
            face_result = {
                "box": face.get("box", []),
                "confidence": face.get("confidence", 0.0)
            }
            
            if data.get("return_landmarks", False):
                face_result["landmarks"] = face.get("landmarks", {})
            
            result["faces"].append(face_result)
        
        logger.info(f"{len(result['faces'])} Gesicht(er) erkannt")
        return jsonify(result)
        
    except Exception as e:
        logger.error(f"Fehler bei Face Detection: {e}", exc_info=True)
        return jsonify({"error": str(e)}), 500


if __name__ == "__main__":
    logger.info("Starte Face Detection Service auf Port 5001")
    logger.info(f"MLX verfügbar: {MLX_AVAILABLE}")
    app.run(host="0.0.0.0", port=5001, debug=False)
