pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mac-studio-cluster"
version = "0.1.0"
description = "Face Tagging System für Mac Studio Cluster mit MLX Integration"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Mac Studio Cluster Team"}
]
keywords = ["mlx", "face-detection", "face-tagging", "apple-silicon", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "mlx>=0.29.0",
    "mlx-lm>=0.28.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "minio>=7.0.0",
    "neo4j>=5.0.0",
    "requests>=2.28.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "ray[default]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
mlx-monitor = "mlx_monitor:main"

[tool.setuptools]
py-modules = ["mlx_monitor"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.html"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

